+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Fixed
+
+- bug in main link for CDS
+- lock tables and CDS
+
 ## 0.4.12 2019-02-06
 
 ### Fixed